Pull top app bar functionality into EntityPage
