+++ visitz/Visitz/Views/Entity/EntityViewModel.cs
@@ -1,17 +1,8 @@
-using System.ComponentModel;
-using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.Input;
-using CommunityToolkit.Mvvm.Messaging;
-using Microsoft.Extensions.Logging;
 using Realms;
 using Syncfusion.Maui.Toolkit.TabView;
 using Visitz.Extensions;
 using Visitz.FontIcons;
-using Visitz.Resources.Localization;
 using Visitz.Resources.Styles;
-using Visitz.Services;
-using Visitz.Services.Base;
-using Visitz.Services.Caseload;
 using Visitz.Storage;
 using Visitz.Views.BaseClasses;
 using Visitz.Views.Entity.Attachments;
@@ -19,11 +10,8 @@ using Visitz.Views.Entity.ChildYouthVisits;
 using Visitz.Views.Entity.Details;
 using Visitz.Views.Entity.Notes;
 using Visitz.Views.Entity.SafetyAssess;
-using VisitzModel.Extensions.EntityTypes;
-using VisitzModel.Messaging;
 using VisitzModel.Models;
 using VisitzModel.Models.Attachments;
-using VisitzModel.Models.Caseload;
 using VisitzModel.Models.Drafts;
 using VisitzModel.Models.InPersonVisits;
 using VisitzModel.Models.Navigation;
@@ -34,28 +22,17 @@ namespace Visitz.Views.Entity;
 
 #nullable enable
 
-public partial class EntityViewModel : IcmRecordViewModel, IRecipient<ServiceStateMessage>
+public partial class EntityViewModel : IcmRecordViewModel
 {
-    ServiceHandler ServiceHandler { get; } = ServiceProvider.GetService<ServiceHandler>();
-
-    string? _cacheRemovedDisplayName;
-
     public EntitySection? RequestedSection { get; set; }
 
     public IDraftItem? FocusedDraftItem { get; set; }
 
     readonly ObservableRealmQueryMap _queryMap = new();
 
-    [ObservableProperty]
-    public bool downloadActivity;
-
     protected override async Task InitAsync()
     {
         await base.InitAsync();
-
-        _cacheRemovedDisplayName = BusinessObject.DisplayName;
-        BusinessObject.SubscribePropertyChanged(BusinessObject_PropertyChanged);
-        WeakReferenceMessenger.Default.Register(this, GetAllDataForRecordService.MakeId(BusinessObject));
 
         try
         {
@@ -65,13 +42,6 @@ public partial class EntityViewModel : IcmRecordViewModel, IRecipient<ServiceSta
                 SelectedTab = GetMappedNavItem(RequestedSection);
 
             SelectedTab ??= GetTabByType<EntityDetailsView>();
-
-            UpdateDownloadActivity();
-
-            ServiceHandler.ServiceStarted += ServiceHandler_ServiceStarted;
-            ServiceHandler.ServiceFinished += ServiceHandler_ServiceFinished;
-
-            UpdateLocalActivityTimestamp();
 
             await SetupDraftIndicatorObservers();
         }
@@ -89,93 +59,9 @@ public partial class EntityViewModel : IcmRecordViewModel, IRecipient<ServiceSta
         {
             DisposeTabViews();
 
-            BusinessObject.UnsubscribePropertyChanged(BusinessObject_PropertyChanged);
-
-            ServiceHandler.ServiceFinished -= ServiceHandler_ServiceFinished;
-            ServiceHandler.ServiceStarted -= ServiceHandler_ServiceStarted;
-
-            WeakReferenceMessenger.Default.UnregisterAll(this);
-
             disposed = true;
         }
         base.Dispose(disposing);
-    }
-
-    async void BusinessObject_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-    {
-        if (sender is not IBusinessObject bobj)
-            return;
-
-        if (e.PropertyName == nameof(bobj.IsValid) && !bobj.IsValid)
-            await EntityUnassignedGoBack();
-    }
-
-    async Task EntityUnassignedGoBack()
-    {
-        GoBack();
-
-        string typeString = EntityType.GetDisplayString();
-
-        await Navigator.CurrentOpenPage.DisplayAlertAsync(
-            string.Format(LocalizedStrings.RecordRemovedFromCaseload, typeString, _cacheRemovedDisplayName),
-            string.Format(LocalizedStrings.RecordRemovedFromCaseloadDetails, typeString, _cacheRemovedDisplayName),
-            LocalizedStrings.Ok
-        );
-    }
-
-    void ServiceHandler_ServiceStarted(object? sender, string e)
-    {
-        try
-        {
-            MainThread.BeginInvokeOnMainThread(UpdateDownloadActivity);
-        }
-        catch (Exception ex)
-        {
-            Logger.LogError(ex, ex.Message);
-        }
-    }
-
-    void ServiceHandler_ServiceFinished(object? sender, VisitzService e)
-    {
-        try
-        {
-            MainThread.BeginInvokeOnMainThread(UpdateDownloadActivity);
-        }
-        catch (Exception ex)
-        {
-            Logger.LogError(ex, ex.Message);
-        }
-    }
-
-    void UpdateDownloadActivity()
-    {
-        DownloadActivity = BusinessObject.IsValid && ServiceHandler.IsAnyServiceRunning(BusinessObject.Id);
-    }
-
-    [RelayCommand]
-    public static void GoBack()
-    {
-        StrongReferenceMessenger.Default.Send(new EntityNavBackMessage());
-    }
-
-    public async void Receive(ServiceStateMessage message)
-    {
-        if (message.FinishedError)
-        {
-            string displayString = $"{EntityType.GetDisplayString()} {_cacheRemovedDisplayName}";
-            string msg = string.Format(LocalizedStrings.DownloadRecordErrorMessage, displayString);
-            await Navigator.CurrentOpenPage.DisplayErrorAlert(
-                msg,
-                message.UncaughtException?.ToString() ?? string.Empty,
-                LocalizedStrings.DownloadError
-            );
-        }
-    }
-
-    void UpdateLocalActivityTimestamp()
-    {
-        if (BusinessObject.IsValid)
-            BusinessObject.LocalState?.LastOpenedBinding = DateTimeOffset.UtcNow;
     }
 
     async Task SetupDraftIndicatorObservers()
